+++ common/src/DataBaseParser.cpp
@@ -14,7 +14,7 @@ void DataBaseParser::put(const std::string &key, const std::string &value)
     _map[key] = value;
 }
 
-void DataBaseParser::put(const std::string &key, const unsigned long &value)
+void DataBaseParser::put(const std::string &key, const unsigned long long &value)
 {
     std::stringstream ss("");
     
@@ -51,7 +51,7 @@ void DataBaseParser::get(const std::string &key, std::string &value)
     value = _map[key];
 }
 
-void DataBaseParser::get(const std::string &key, unsigned long &value)
+void DataBaseParser::get(const std::string &key, unsigned long long &value)
 {
     std::string str = _map[key];
     std::stringstream ss(str);
